Dashboard Modifications: Consolidated Dashboard View, Separated Admin controls, various aesthetic improvements
