+++ src/App.tsx
@@ -17,6 +17,7 @@ import InternalSettings from "./components/internalSettings/InternalSettings";
 import CheckValidHexathon from "./util/CheckValidHexathon";
 import UserInfoTable from "./components/userInfo/UserInfoTable";
 import "./App.css";
+import AdminControlsHome from "./components/dashboard/AdminControlsHome";
 
 axios.defaults.withCredentials = true;
 
@@ -36,6 +37,7 @@ export const App = () => {
           ) : (
             <Routes>
               <Route path="/" element={<SelectEvent />} />
+              <Route path="/admin" element={<AdminControlsHome/>} />
               <Route path="/:hexathonId" element={<CheckValidHexathon />}>
                 <Route path="" element={<Dashboard />} />
                 <Route path="user/:userId" element={<User />} />
